agents/5-notion-agent/requirements.txt
jupyterlab
ipykernel
ipywidgets
openai-agents
openai
python-dotenv